Fix w-otp DIGITS=10 modulus divergence; close parity-gate and surface gaps
- go, rust: compute the w-otp modulus in 64 bits. CRYPTO_SPEC mandates
  otp = value mod 10^DIGITS; Go's uint32 modulus wrapped 10^10 to
  1410065408 (divergent codes for ~2/3 of HMAC values at DIGITS=10) and
  Rust's saturating u32 clamped it to 2^32-1.
- tools/check_wotp_parity.sh: sweep DIGITS 4/6/10 with a probe WID whose
  32-bit HMAC word (0xf960d3a5) makes wrap/saturation bugs
  deterministically visible instead of hash-dependent.
- sh: fail loudly when the uv delegation fails (a failed `if` condition
  returned success with no output under set -e).
- c: drop the dead has_unsafe_shell_char filter — the C CLI never shells
  out, and the filter made C alone reject KEY/DATA values the other five
  implementations accept.
- python, typescript, README: scope the shared-sqlite "no duplicate WIDs"
  guarantee to the CLI E=sql CAS path; the library stores are documented
  as single-process last-writer-wins.
- README: disclose that `make next` (I=auto) delegates to the Python
  implementation when present.
- dist/: rebuilt (committed-dist policy; CI freshness gate).

+++ go/cmd/wid/main.go
@@ -682,8 +682,10 @@ func computeWOtp(secret, widValue string, digits int) string {
 	mac := hmac.New(sha256.New, []byte(secret))
 	_, _ = mac.Write([]byte(widValue))
 	sum := mac.Sum(nil)
-	v := (uint32(sum[0]) << 24) | (uint32(sum[1]) << 16) | (uint32(sum[2]) << 8) | uint32(sum[3])
-	mod := uint32(1)
+	v := (uint64(sum[0]) << 24) | (uint64(sum[1]) << 16) | (uint64(sum[2]) << 8) | uint64(sum[3])
+	// CRYPTO_SPEC: otp = value mod 10^DIGITS. The modulus must be 64-bit:
+	// DIGITS may be 10 and 10^10 wraps a uint32 (10^10 mod 2^32 = 1410065408).
+	mod := uint64(1)
 	for i := 0; i < digits; i++ {
 		mod *= 10
 	}
